requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Evaluation & Statistics
scipy>=1.11.0
nltk>=3.8.0
rouge-score>=0.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0  # Optional: for experiment tracking
pyyaml>=6.0
jsonlines>=3.1.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
